Merge dev with main

+++ Services/items.js
@@ -1,46 +1,62 @@
-import api from "./apiConfig"
+import api from "./apiConfig";
 
 export const getItems = async () => {
   try {
-    const response = await api.get('/items')
-    return response.data
+    const response = await api.get("/items");
+    return response.data;
   } catch (error) {
-    console.error(error)
+    console.error(error);
   }
-}
+};
 
 export const getItem = async (itemId) => {
   try {
-    const response = await api.get(`/items/${itemId}`)
-    return response.data
+    const response = await api.get(`/items/${itemId}`);
+    return response.data;
   } catch (error) {
-    console.error(error)
-  }
-}
-
-export const createItem = async (itemData) => {
-  try {
-    const response = await api.post('/items', itemData)
-    return response.data
-  } catch (error) {
-    console.error(error)
+    console.error(error);
   }
 };
 
-export const updateItem = async (itemId, itemData) => {
+export const createItem = async (itemData, userId) => {
   try {
-    const response = await api.put(`/items/${itemId}`, itemData)
-    return response.data
+    const token = localStorage.getItem("token");
+    const response = await api.post(`/items/${userId}`, itemData, {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    });
+    console.log("Items working");
+    return response.data;
   } catch (error) {
-    console.error(error)
+    console.error(error);
   }
 };
 
-export const deleteItem = async (itemId) => {
+export const updateItem = async (userId, itemId, itemData) => {
   try {
-    const response = await api.delete(`/items/${itemId}`)
-    return response.data
+    const token = localStorage.getItem("token");
+    const response = await api.put(`/items/${userId}/${itemId}`, itemData, {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    });
+    return response.data;
   } catch (error) {
-    console.error(error)
+    console.error(error);
   }
-}
+};
+
+export const deleteItem = async (userId,itemId) => {
+  try {
+    const token = localStorage.getItem("token");
+    const response = await api.delete(`/items/${userId}/${itemId}`, {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    });
+    return response.data;
+  } catch (error) {
+    console.error(error);
+  }
+};
